requirements.txt
pandas
streamlit
google-genai
faiss-cpu
numpy
python-dotenv
sentence-transformers
tqdm
